Fix exit play mode process for domain reloading is enabled
