+++ Runtime/Launcher.cs
@@ -247,7 +247,7 @@ namespace DeNA.Anjin
                 await UniTask.NextFrame();
 #if UNITY_EDITOR
                 EditorApplication.isPlaying = false;
-                // Note: If launched from the command line, `DeNA.Anjin.Editor.Commandline.OnChangePlayModeState()` will be called, and the Unity editor will be terminated.
+                // Note: If launched from the command line, `DeNA.Anjin.Autopilot.OnExitPlayModeToTerminateEditor()` will be called, and the Unity editor will be terminated.
 #endif
             }
             else
